+++ packages/stargate/src/modules/feegrant/aminomessages.ts
@@ -1,6 +1,6 @@
 import { AminoConverters } from "../../aminotypes";
 
-export function createFreegrantAminoConverters(): AminoConverters {
+export function createFeegrantAminoConverters(): AminoConverters {
   return {
     "/cosmos.feegrant.v1beta1.MsgGrantAllowance": "not_supported_by_chain",
     "/cosmos.feegrant.v1beta1.MsgRevokeAllowance": "not_supported_by_chain",