+++ packages/stargate/src/modules/index.ts
@@ -38,7 +38,7 @@ export {
   createEvidenceAminoConverters,
   isAminoMsgSubmitEvidence,
 } from "./evidence/aminomessages";
-export { createFreegrantAminoConverters } from "./feegrant/aminomessages";
+export { createFeegrantAminoConverters } from "./feegrant/aminomessages";
 export { feegrantTypes } from "./feegrant/messages";
 export { FeegrantExtension, setupFeegrantExtension } from "./feegrant/queries";
 export {