+++ CHANGELOG.md
@@ -76,6 +76,8 @@ and this project adheres to
   pagination key ([#1095]).
 - @cosmjs/proto-signing: `makeAuthInfoBytes` now expects a fee granter and fee
   payer argument in position 4 and 5.
+- @cosmjs/stargate: Rename exported function `createFreegrantAminoConverters` to
+  `createFeegrantAminoConverters` due to a typo ([#1261).
 
 [#1131]: https://github.com/cosmos/cosmjs/pull/1131
 [#1168]: https://github.com/cosmos/cosmjs/pull/1168
@@ -84,6 +86,7 @@ and this project adheres to
 [#1234]: https://github.com/cosmos/cosmjs/issues/1234
 [#1095]: https://github.com/cosmos/cosmjs/issues/1095
 [#1254]: https://github.com/cosmos/cosmjs/issues/1254
+[#1261]: https://github.com/cosmos/cosmjs/pull/1261
 
 ## [0.28.11] - 2022-07-13